requirements.txt
alembic==1.4.2
beautifulsoup4==4.9.0
Bootstrap-Flask==1.3.1
certifi==2020.4.5.1
chardet==3.0.4
Click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jsonify==0.5
Mako==1.1.2
MarkupSafe==1.1.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.14.0
soupsieve
SQLAlchemy==1.3.16
sqlalchemy-datatables==2.0.1
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
psycopg2
APScheduler==3.6.3
flask-socketio
python-socketio
gevent-websocket
gevent
flask-cors
mysqlclient
flask-sqlacodegen
Flask-User
bcrypt
Flask-Login
Flask-Babel
WTForms-Alchemy
holidays
Flask-Migrate
PyMySQL
configparser==5.0.0
Flask-Minio
Flask-SQLAlchemy-Caching
email_validator